에러 안나게만 수정
